Fix OpenHours import
When updating this gem’s version to 1.0 on a project, I kept running
into the same issue:

  cannot load such file -- yelp/fusion/responses/models/openHours

After a quick search in the gem’s source code, I’ve noticed some
unexpected capitalisation that I was not expecting. Thanks to it,
we were basically trying to load the wrong file. Changing it to
snake_case fixed the loading issue for my app.